+++ frontend/src/components/AddForm.jsx
@@ -61,7 +61,7 @@ const AddForm = ({ content, date, onSuccess }) => {
           credentials: 'same-origin',
           headers: {
             'Content-Type': 'application/json',
-            'x-app-token': token,
+            'X-App-Token': token,
           },
           redirect: 'follow',
           referrerPolicy: 'no-referrer',