+++ frontend/src/views/OneDay.jsx
@@ -150,7 +150,7 @@ const OneDay = () => {
           credentials: 'same-origin',
           headers: {
             'Content-Type': 'application/json',
-            'x-app-token': token,
+            'X-App-Token': token,
           },
           redirect: 'follow',
           referrerPolicy: 'no-referrer',